refactor(engine): commands.rs 按引擎臂拆四个平铺文件
- 5432→1340 行：send 2553（8 臂自由函数+调度器）/send_sync 1121/opencode_catalog 693/pi_rpc 321
- 逐字搬运；super:: 路径改写 crate::engine::；唯一 pub(super) 升 pub(crate)
- engine/mod.rs 与 command_registry 零改动（pub use glob 保持）
- Dual-Path Gate 合规：未触碰转发器/门控/结算语义，daemon 侧零改动
- baseline 重生成 + 新文件登记豁免
验证：cargo check 零 error；engine 测试失败集 ⊆ 基线 14 个既有失败（零新增）；rustfmt 过；gate 绿
（子代理 CommandsSplit 执行，主代理复核验证）
