+++ src-tauri/src/engine/commands_opencode_catalog.rs
@@ -1,0 +1,693 @@
+//! OpenCode catalog types and provider/session helper functions, split from `commands.rs`.
+
+use super::*;
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OpenCodeCommandEntry {
+    pub name: String,
+    pub description: Option<String>,
+    #[serde(rename = "argumentHint")]
+    pub argument_hint: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OpenCodeAgentEntry {
+    pub id: String,
+    pub description: Option<String>,
+    pub is_primary: bool,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OpenCodeProviderHealth {
+    pub provider: String,
+    pub connected: bool,
+    pub credential_count: usize,
+    pub matched: bool,
+    pub authenticated_providers: Vec<String>,
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OpenCodeMcpServerState {
+    pub name: String,
+    pub enabled: bool,
+    pub status: Option<String>,
+    pub permission_hint: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OpenCodeStatusSnapshot {
+    pub session_id: Option<String>,
+    pub model: Option<String>,
+    pub agent: Option<String>,
+    pub variant: Option<String>,
+    pub provider: Option<String>,
+    pub provider_health: OpenCodeProviderHealth,
+    pub mcp_enabled: bool,
+    pub mcp_servers: Vec<OpenCodeMcpServerState>,
+    pub mcp_raw: String,
+    pub managed_toggles: bool,
+    pub token_usage: Option<u64>,
+    pub context_window: Option<u64>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OpenCodeSessionEntry {
+    pub session_id: String,
+    pub title: String,
+    pub updated_label: String,
+    pub updated_at: Option<i64>,
+    /// Session working directory from OpenCode (`session list --format json`).
+    /// Used to filter out global/foreign project leakage into empty workspaces.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub directory: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OpenCodeProviderOption {
+    pub id: String,
+    pub label: String,
+    pub description: Option<String>,
+    pub category: String,
+    pub recommended: bool,
+}
+
+#[derive(Debug, Clone, Default)]
+pub(crate) struct OpenCodeMcpToggleState {
+    pub(crate) global_enabled: bool,
+    pub(crate) server_enabled: HashMap<String, bool>,
+}
+
+pub(crate) const OPENCODE_CACHE_TTL: Duration = Duration::from_secs(30);
+pub(crate) static OPENCODE_COMMANDS_CACHE: OnceLock<
+    Mutex<Option<(Instant, Vec<OpenCodeCommandEntry>)>>,
+> = OnceLock::new();
+pub(crate) static OPENCODE_AGENTS_CACHE: OnceLock<
+    Mutex<Option<(Instant, Vec<OpenCodeAgentEntry>)>>,
+> = OnceLock::new();
+pub(crate) static OPENCODE_MCP_TOGGLE_STATE: OnceLock<
+    Mutex<HashMap<String, OpenCodeMcpToggleState>>,
+> = OnceLock::new();
+
+pub(crate) fn strip_ansi_codes(input: &str) -> String {
+    let mut out = String::with_capacity(input.len());
+    let mut chars = input.chars().peekable();
+    while let Some(ch) = chars.next() {
+        if ch == '\u{1b}' {
+            if let Some('[') = chars.peek().copied() {
+                let _ = chars.next();
+                for c in chars.by_ref() {
+                    if ('@'..='~').contains(&c) {
+                        break;
+                    }
+                }
+                continue;
+            }
+        }
+        out.push(ch);
+    }
+    out
+}
+
+pub(crate) fn extract_turn_result_text_internal(value: &Value, depth: usize) -> Option<String> {
+    if depth > 4 {
+        return None;
+    }
+    if let Some(text) = value
+        .as_str()
+        .map(str::trim)
+        .filter(|text| !text.is_empty())
+    {
+        return Some(text.to_string());
+    }
+    if let Some(array) = value.as_array() {
+        let mut merged = String::new();
+        for item in array {
+            if let Some(text) = extract_turn_result_text_internal(item, depth + 1) {
+                if !merged.is_empty() {
+                    merged.push('\n');
+                }
+                merged.push_str(&text);
+            }
+        }
+        return if merged.trim().is_empty() {
+            None
+        } else {
+            Some(merged)
+        };
+    }
+    if let Some(object) = value.as_object() {
+        for key in [
+            "text",
+            "delta",
+            "output_text",
+            "outputText",
+            "content",
+            "message",
+        ] {
+            if let Some(text) = object
+                .get(key)
+                .and_then(|entry| entry.as_str())
+                .map(str::trim)
+                .filter(|entry| !entry.is_empty())
+            {
+                return Some(text.to_string());
+            }
+        }
+        for key in [
+            "result", "response", "content", "message", "output", "data", "payload",
+        ] {
+            if let Some(entry) = object.get(key) {
+                if let Some(text) = extract_turn_result_text_internal(entry, depth + 1) {
+                    return Some(text);
+                }
+            }
+        }
+    }
+    None
+}
+
+pub(crate) fn extract_turn_result_text(result: Option<&Value>) -> Option<String> {
+    result.and_then(|value| extract_turn_result_text_internal(value, 0))
+}
+
+pub(crate) fn should_prefer_turn_result_text(result: Option<&Value>) -> bool {
+    result
+        .and_then(|value| value.get("syntheticApprovalResolved"))
+        .and_then(Value::as_bool)
+        .unwrap_or(false)
+}
+
+pub(crate) fn is_likely_foreign_model_for_gemini(model: &str) -> bool {
+    let normalized = model.trim().to_ascii_lowercase();
+    if normalized.is_empty() {
+        return false;
+    }
+    if normalized.contains("gemini") {
+        return false;
+    }
+    if normalized.starts_with("claude-") {
+        return true;
+    }
+    if normalized.starts_with("gpt-") || normalized.contains("codex") {
+        return true;
+    }
+    normalized.starts_with("openai/")
+        || normalized.starts_with("anthropic/")
+        || normalized.starts_with("x-ai/")
+        || normalized.starts_with("openrouter/")
+        || normalized.starts_with("deepseek/")
+        || normalized.starts_with("qwen/")
+        || normalized.starts_with("meta/")
+        || normalized.starts_with("mistral/")
+}
+
+pub(crate) fn is_likely_legacy_claude_model_id(model: &str) -> bool {
+    model.trim().to_ascii_lowercase().starts_with("claude-")
+}
+
+pub(crate) fn is_valid_claude_model_for_passthrough(model: &str) -> bool {
+    let trimmed = model.trim();
+    if trimmed.is_empty() || trimmed.len() > 128 {
+        return false;
+    }
+    trimmed.chars().all(|ch| {
+        ch.is_ascii_alphanumeric() || matches!(ch, '-' | '_' | '.' | ':' | '/' | '[' | ']')
+    })
+}
+
+pub(crate) fn resolve_opencode_bin(config: Option<&EngineConfig>) -> Result<String, String> {
+    let custom_bin = config.and_then(|c| c.bin_path.as_deref());
+    crate::backend::app_server_cli::resolve_safe_opencode_binary(custom_bin)
+        .map(|path| path.to_string_lossy().to_string())
+}
+
+pub(crate) fn build_opencode_command(
+    config: Option<&EngineConfig>,
+) -> Result<crate::engine::opencode_native_artifact::ContainedOpenCodeCommand, String> {
+    let bin = resolve_opencode_bin(config)?;
+    let mut cmd = crate::backend::app_server::build_command_for_binary(&bin);
+    if let Some(home) = config.and_then(|c| c.home_dir.as_ref()) {
+        cmd.env("OPENCODE_HOME", home);
+    }
+    crate::engine::opencode_native_artifact::ContainedOpenCodeCommand::new(cmd)
+}
+
+pub(crate) fn opencode_session_candidate_paths(
+    workspace_path: &Path,
+    session_id: &str,
+    config: Option<&EngineConfig>,
+) -> Vec<PathBuf> {
+    let mut roots = Vec::new();
+    if let Some(home) = config.and_then(|item| item.home_dir.as_ref()) {
+        roots.push(PathBuf::from(home).join("sessions"));
+    }
+    if let Some(home) = std::env::var_os("OPENCODE_HOME") {
+        roots.push(PathBuf::from(home).join("sessions"));
+    }
+    if let Some(home) = dirs::home_dir() {
+        roots.push(home.join(".opencode").join("sessions"));
+    }
+    roots.push(workspace_path.join(".opencode").join("sessions"));
+
+    let mut candidates = Vec::new();
+    for root in roots {
+        for candidate in [
+            root.join(session_id),
+            root.join(format!("{session_id}.json")),
+        ] {
+            if !candidates.contains(&candidate) {
+                candidates.push(candidate);
+            }
+        }
+    }
+    candidates
+}
+
+pub(crate) fn delete_opencode_session_files(
+    workspace_path: &Path,
+    session_id: &str,
+    config: Option<&EngineConfig>,
+) -> Result<(), String> {
+    let normalized_session_id = session_id.trim();
+    if normalized_session_id.is_empty()
+        || normalized_session_id.contains('/')
+        || normalized_session_id.contains('\\')
+        || normalized_session_id.contains("..")
+    {
+        return Err("[SESSION_NOT_FOUND] Invalid OpenCode session id".to_string());
+    }
+
+    let mut deleted_any = false;
+
+    let candidates =
+        opencode_session_candidate_paths(workspace_path, normalized_session_id, config);
+    for candidate in candidates {
+        if !candidate.exists() {
+            continue;
+        }
+        let delete_result = if candidate.is_dir() {
+            fs::remove_dir_all(&candidate)
+        } else {
+            fs::remove_file(&candidate)
+        };
+        match delete_result {
+            Ok(()) => {
+                deleted_any = true;
+            }
+            Err(error) if error.kind() == std::io::ErrorKind::NotFound => continue,
+            Err(error) => {
+                return Err(format!(
+                    "[IO_ERROR] Failed to delete OpenCode session path {}: {}",
+                    candidate.display(),
+                    error
+                ));
+            }
+        }
+    }
+
+    for data_root in opencode_data_candidate_roots(workspace_path, config) {
+        match delete_opencode_session_from_datastore(&data_root, normalized_session_id) {
+            Ok(true) => {
+                deleted_any = true;
+            }
+            Ok(false) => {}
+            Err(error) => return Err(error),
+        }
+    }
+
+    if deleted_any {
+        return Ok(());
+    }
+
+    Err(format!(
+        "[SESSION_NOT_FOUND] OpenCode session file not found: {}",
+        normalized_session_id
+    ))
+}
+
+pub(crate) fn opencode_data_candidate_roots(
+    workspace_path: &Path,
+    config: Option<&EngineConfig>,
+) -> Vec<PathBuf> {
+    let mut roots = Vec::new();
+    if let Some(home) = config.and_then(|item| item.home_dir.as_ref()) {
+        roots.push(PathBuf::from(home));
+    }
+    if let Some(home) = std::env::var_os("OPENCODE_HOME") {
+        roots.push(PathBuf::from(home));
+    }
+    if let Some(data_home) = dirs::data_local_dir() {
+        roots.push(data_home.join("opencode"));
+    }
+    if let Some(data_dir) = dirs::data_dir() {
+        roots.push(data_dir.join("opencode"));
+    }
+    if let Some(home) = dirs::home_dir() {
+        roots.push(home.join(".local").join("share").join("opencode"));
+    }
+    roots.push(workspace_path.join(".opencode"));
+
+    let mut deduped = Vec::new();
+    for root in roots {
+        if !deduped.contains(&root) {
+            deduped.push(root);
+        }
+    }
+    deduped
+}
+
+pub(crate) fn delete_path_if_exists(path: &Path) -> Result<bool, String> {
+    if !path.exists() {
+        return Ok(false);
+    }
+    let result = if path.is_dir() {
+        fs::remove_dir_all(path)
+    } else {
+        fs::remove_file(path)
+    };
+    match result {
+        Ok(()) => Ok(true),
+        Err(error) if error.kind() == std::io::ErrorKind::NotFound => Ok(false),
+        Err(error) => Err(format!(
+            "[IO_ERROR] Failed to delete OpenCode session path {}: {}",
+            path.display(),
+            error
+        )),
+    }
+}
+
+pub(crate) fn delete_opencode_session_from_datastore(
+    data_root: &Path,
+    session_id: &str,
+) -> Result<bool, String> {
+    let mut deleted_any = false;
+
+    let db_path = data_root.join("opencode.db");
+    if db_path.exists() {
+        let connection = Connection::open(&db_path).map_err(|error| {
+            format!(
+                "[IO_ERROR] Failed to open OpenCode datastore {}: {}",
+                db_path.display(),
+                error
+            )
+        })?;
+        connection
+            .pragma_update(None, "foreign_keys", "ON")
+            .map_err(|error| {
+                format!(
+                    "[IO_ERROR] Failed to enable OpenCode datastore foreign_keys {}: {}",
+                    db_path.display(),
+                    error
+                )
+            })?;
+        let deleted_rows = connection
+            .execute("DELETE FROM session WHERE id = ?1", params![session_id])
+            .map_err(|error| {
+                format!(
+                    "[IO_ERROR] Failed to delete OpenCode session {} in {}: {}",
+                    session_id,
+                    db_path.display(),
+                    error
+                )
+            })?;
+        if deleted_rows > 0 {
+            deleted_any = true;
+        }
+    }
+
+    let storage_root = data_root.join("storage");
+    if storage_root.exists() {
+        let reader = fs::read_dir(&storage_root).map_err(|error| {
+            format!(
+                "[IO_ERROR] Failed to read OpenCode storage directory {}: {}",
+                storage_root.display(),
+                error
+            )
+        })?;
+        for entry in reader {
+            let entry = entry.map_err(|error| {
+                format!(
+                    "[IO_ERROR] Failed to read OpenCode storage entry under {}: {}",
+                    storage_root.display(),
+                    error
+                )
+            })?;
+            let parent = entry.path();
+            if !parent.is_dir() {
+                continue;
+            }
+            if delete_path_if_exists(&parent.join(session_id))? {
+                deleted_any = true;
+            }
+            if delete_path_if_exists(&parent.join(format!("{session_id}.json")))? {
+                deleted_any = true;
+            }
+        }
+    }
+
+    Ok(deleted_any)
+}
+
+pub(crate) fn slugify_provider_label(value: &str) -> String {
+    let mut out = String::new();
+    let mut last_dash = false;
+    for ch in value.chars() {
+        if ch.is_ascii_alphanumeric() {
+            out.push(ch.to_ascii_lowercase());
+            last_dash = false;
+            continue;
+        }
+        if !last_dash {
+            out.push('-');
+            last_dash = true;
+        }
+    }
+    out.trim_matches('-').to_string()
+}
+
+pub(crate) fn parse_provider_option_line(
+    line: &str,
+    category: &str,
+) -> Option<OpenCodeProviderOption> {
+    let trimmed = line
+        .trim_start_matches(|ch: char| matches!(ch, '●' | '○' | '◆' | '◇' | '│'))
+        .trim();
+    if trimmed.is_empty() || trimmed.starts_with("Search:") || trimmed == "..." {
+        return None;
+    }
+    let lower = trimmed.to_ascii_lowercase();
+    if lower == "select provider"
+        || lower == "add credential"
+        || lower == "login method"
+        || lower.contains("to select")
+        || lower.contains("enter: confirm")
+        || lower.contains("type: to search")
+        || lower.starts_with("search:")
+        || trimmed.starts_with('┌')
+        || trimmed.starts_with('└')
+        || trimmed.starts_with('■')
+        || trimmed.starts_with('│')
+    {
+        return None;
+    }
+    let (label, description) = if let Some((left, right)) = trimmed.split_once('(') {
+        (
+            left.trim().to_string(),
+            Some(right.trim_end_matches(')').trim().to_string()),
+        )
+    } else {
+        (trimmed.to_string(), None)
+    };
+    if label.is_empty() {
+        return None;
+    }
+    let id = slugify_provider_label(&label);
+    if id.is_empty() {
+        return None;
+    }
+    let recommended = description
+        .as_ref()
+        .map(|text| text.to_ascii_lowercase().contains("recommended"))
+        .unwrap_or(false);
+    Some(OpenCodeProviderOption {
+        id,
+        label,
+        description,
+        category: category.to_string(),
+        recommended,
+    })
+}
+
+pub(crate) fn fallback_opencode_provider_catalog() -> Vec<OpenCodeProviderOption> {
+    let popular = vec![
+        ("opencode-zen", "OpenCode Zen", Some("recommended")),
+        ("anthropic", "Anthropic", Some("Claude Max or API key")),
+        ("github-copilot", "GitHub Copilot", None),
+        ("openai", "OpenAI", Some("ChatGPT Plus/Pro or API key")),
+        ("google", "Google", None),
+    ];
+    let other = vec![
+        ("z-ai", "Z.AI"),
+        ("zenmux", "ZenMux"),
+        ("io-net", "IO.NET"),
+        ("nvidia", "Nvidia"),
+        ("fastrouter", "FastRouter"),
+        ("iflow", "iFlow"),
+        ("modelscope", "ModelScope"),
+        ("llama", "Llama"),
+    ];
+
+    let mut out = Vec::new();
+    for (id, label, description) in popular {
+        out.push(OpenCodeProviderOption {
+            id: id.to_string(),
+            label: label.to_string(),
+            description: description.map(ToOwned::to_owned),
+            category: "popular".to_string(),
+            recommended: description
+                .map(|text| text.to_ascii_lowercase().contains("recommended"))
+                .unwrap_or(false),
+        });
+    }
+    for (id, label) in other {
+        out.push(OpenCodeProviderOption {
+            id: id.to_string(),
+            label: label.to_string(),
+            description: None,
+            category: "other".to_string(),
+            recommended: false,
+        });
+    }
+    out
+}
+
+pub(crate) async fn fetch_opencode_provider_catalog_preview(
+    workspace_path: &PathBuf,
+    config: Option<&EngineConfig>,
+) -> Vec<OpenCodeProviderOption> {
+    let mut cmd = match build_opencode_command(config) {
+        Ok(value) => value,
+        Err(_) => return Vec::new(),
+    };
+    cmd.current_dir(workspace_path);
+    cmd.arg("auth");
+    cmd.arg("login");
+    cmd.stdout(Stdio::piped());
+    cmd.stderr(Stdio::piped());
+    cmd.stdin(Stdio::null());
+    let mut child = match cmd.spawn() {
+        Ok(value) => value,
+        Err(_) => return Vec::new(),
+    };
+    tokio::time::sleep(Duration::from_millis(900)).await;
+    let _ = child.start_kill();
+    let output = match tokio::time::timeout(Duration::from_secs(2), child.wait_with_output()).await
+    {
+        Ok(Ok(value)) => value,
+        _ => return Vec::new(),
+    };
+    let stdout = strip_ansi_codes(&String::from_utf8_lossy(&output.stdout));
+    let mut providers: Vec<OpenCodeProviderOption> = Vec::new();
+    let mut category = "popular".to_string();
+    for line in stdout.lines() {
+        let trimmed = line.trim();
+        if trimmed.eq_ignore_ascii_case("Popular") {
+            category = "popular".to_string();
+            continue;
+        }
+        if trimmed.eq_ignore_ascii_case("Other") {
+            category = "other".to_string();
+            continue;
+        }
+        if let Some(option) = parse_provider_option_line(line, &category) {
+            providers.push(option);
+        }
+    }
+    providers.sort_by(|a, b| a.label.cmp(&b.label));
+    providers.dedup_by(|a, b| a.id == b.id);
+    providers
+}
+
+pub(crate) async fn fetch_opencode_provider_catalog_from_auth_picker(
+    workspace_path: &PathBuf,
+    config: Option<&EngineConfig>,
+) -> Vec<OpenCodeProviderOption> {
+    let mut cmd = match build_opencode_command(config) {
+        Ok(value) => value,
+        Err(_) => return Vec::new(),
+    };
+    cmd.current_dir(workspace_path);
+    cmd.arg("auth");
+    cmd.arg("login");
+    cmd.stdout(Stdio::piped());
+    cmd.stderr(Stdio::piped());
+    cmd.stdin(Stdio::piped());
+    let mut child = match cmd.spawn() {
+        Ok(value) => value,
+        Err(_) => return Vec::new(),
+    };
+
+    if let Some(stdin) = child.stdin.as_mut() {
+        let mut payload = String::new();
+        for _ in 0..520 {
+            payload.push_str("\u{1b}[B");
+        }
+        payload.push('\u{3}');
+        if stdin.write_all(payload.as_bytes()).await.is_err() {
+            let _ = child.start_kill();
+            return Vec::new();
+        }
+        let _ = stdin.flush().await;
+    }
+
+    let output = match tokio::time::timeout(Duration::from_secs(12), child.wait_with_output()).await
+    {
+        Ok(Ok(value)) => value,
+        _ => return Vec::new(),
+    };
+    let stdout = strip_ansi_codes(&String::from_utf8_lossy(&output.stdout));
+    let mut providers: Vec<OpenCodeProviderOption> = Vec::new();
+    let mut category = "popular".to_string();
+    for line in stdout.lines() {
+        let trimmed = line.trim();
+        if trimmed.eq_ignore_ascii_case("Popular") {
+            category = "popular".to_string();
+            continue;
+        }
+        if trimmed.eq_ignore_ascii_case("Other") {
+            category = "other".to_string();
+            continue;
+        }
+        if let Some(option) = parse_provider_option_line(line, &category) {
+            if let Some(existing) = providers.iter_mut().find(|item| item.id == option.id) {
+                if option.category == "popular" {
+                    existing.category = "popular".to_string();
+                }
+                if existing.description.is_none() && option.description.is_some() {
+                    existing.description = option.description.clone();
+                }
+                existing.recommended = existing.recommended || option.recommended;
+                continue;
+            }
+            providers.push(option);
+        }
+    }
+    providers.sort_by(|a, b| {
+        let score_a = if a.category == "popular" { 0 } else { 1 };
+        let score_b = if b.category == "popular" { 0 } else { 1 };
+        score_a
+            .cmp(&score_b)
+            .then_with(|| b.recommended.cmp(&a.recommended))
+            .then_with(|| a.label.cmp(&b.label))
+    });
+    providers.dedup_by(|a, b| a.id == b.id);
+    providers
+}
